requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aioredis==2.0.1
aiosignal==1.3.2
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.0
beautifulsoup4==4.13.4
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.3
curl_cffi==0.11.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.6
filelock==3.18.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.5.1
google-adk==0.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.94.0
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.16.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.1.1
groq==0.25.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
hf-xet==1.1.2
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-aiohttp==0.1.4
httpx-sse==0.4.0
huggingface-hub==0.32.0
idna==3.10
importlib_metadata==8.6.1
jaraco.context==6.0.1
Jinja2==3.1.6
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.4.2
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-groq==0.3.2
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langsmith==0.3.42
litellm==1.71.1
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.1
more-itertools==10.7.0
multidict==6.4.4
multitasking==0.0.11
mypy_extensions==1.1.0
neo4j==5.28.1
numpy==2.2.6
openai==1.82.0
opentelemetry-api==1.33.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
peewee==3.18.1
platformdirs==4.3.8
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
rq==2.1.0
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.37
sse-starlette==2.3.5
starlette==0.41.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wolframalpha==5.1.3
wrapt==1.17.2
xmltodict==0.14.2
yarl==1.20.0
yfinance==0.2.61
zipp==3.21.0
zstandard==0.23.0
